// server/upstream.generated.ts
export interface UpstreamAudioBlock {
  audio_room_id: number;
  color: string;
  id: number;
  mute_on_entry: boolean;
  name: null | string;
  pos: { x: number; y: number };
  type: "AudioBlock";
}

export interface UpstreamAudioRoom {
  expires_at: null;
  has_audio_block: boolean;
  height: number;
  id: number;
  mute_on_entry: boolean;
  pos: { x: number; y: number };
  type: "AudioRoom";
  width: number;
}

export interface UpstreamAvatar {
  direction: string;
  flair: null | string;
  id: number;
  image_path: string;
  last_idle_at: string;
  last_present_at: string;
  last_seen_at: string;
  message: null | {
    mentioned_entity_ids: Array<number>;
    sent_at: string;
    text: string;
  };
  muted: boolean;
  person_name: string;
  pos: { x: number; y: number };
  rc_hub_visit_notes: null | string;
  rc_hub_visit_today: boolean | null;
  type: "Avatar";
  updated_at: string;
  user_id: number;
  zoom_meeting_topic: null;
  zoom_user_display_name: null;
  zoom_user_id: null;
}

export interface UpstreamBot {
  app: { id: number; name: string };
  can_be_mentioned: boolean;
  direction: string;
  display_name: string;
  emoji: string;
  id: number;
  message: null | {
    mentioned_entity_ids: Array<number>;
    sent_at: string;
    text: string;
  };
  name: string;
  pos: { x: number; y: number };
  type: "Bot";
}

export interface UpstreamDesk {
  color: string;
  emoji: null | string;
  expires_at: null | string;
  id: number;
  owner?: { id: number; image_url: string; name: string };
  pos: { x: number; y: number };
  profile_url: null | string;
  status: null | string;
  type: "Desk";
}

export interface UpstreamLink {
  color: string;
  id: number;
  pos: { x: number; y: number };
  type: "Link";
  url: null | string;
}

export interface UpstreamNote {
  color: string;
  id: number;
  note_text: null | string;
  note_updated_at: null | string;
  pos: { x: number; y: number };
  type: "Note";
  updated_by?: { id: null | number; name: string };
}

export interface UpstreamPhotoBlock {
  color: string;
  id: number;
  photo_attached: boolean;
  pos: { x: number; y: number };
  type: "PhotoBlock";
}

export interface UpstreamRCCalendar {
  calendar_data: Array<{
    end_time: string;
    start_time: string;
    title: string;
    url: string;
  }>;
  color: string;
  id: number;
  pos: { x: number; y: number };
  type: "RC::Calendar";
  updated_at: string;
  zoom_user: { display_name: string; id: number };
}

export interface UpstreamWall {
  color: string;
  id: number;
  pos: { x: number; y: number };
  type: "Wall";
  wall_text: null | string;
}

export interface UpstreamZoomLink {
  color: string;
  id: number;
  participant_count: number;
  pos: { x: number; y: number };
  type: "ZoomLink";
  zoom_user?: { display_name: string; id: number };
}

export type UpstreamEntity =
  | UpstreamAudioBlock
  | UpstreamAudioRoom
  | UpstreamAvatar
  | UpstreamBot
  | UpstreamDesk
  | UpstreamLink
  | UpstreamNote
  | UpstreamPhotoBlock
  | UpstreamRCCalendar
  | UpstreamWall
  | UpstreamZoomLink;